move radius types to rp
