+++ pkg/armrpc/api/v1/types.go
@@ -8,8 +8,6 @@ package v1
 import (
 	"net/http"
 	"strings"
-
-	"github.com/project-radius/radius/pkg/rp/outputresource"
 )
 
 const (
@@ -168,49 +166,4 @@ func (b *BaseResource) ProvisioningState() ProvisioningState {
 // SetProvisioningState sets the privisioning state of the resource.
 func (b *BaseResource) SetProvisioningState(state ProvisioningState) {
 	b.InternalMetadata.AsyncProvisioningState = state
-}
-
-// BasicResourceProperties is the basic resource model for radius resources.
-type BasicResourceProperties struct {
-	// Environment represents the id of environment resource.
-	Environment string `json:"environment,omitempty"`
-	// Application represents the id of application resource.
-	Application string `json:"application,omitempty"`
-
-	// Status represents the resource status.
-	Status ResourceStatus `json:"status,omitempty"`
-}
-
-// BasicDaprResourceProperties is the basic resource properties for dapr resources.
-type BasicDaprResourceProperties struct {
-	// ComponentName represents the name of the component.
-	ComponentName string `json:"componentName,omitempty"`
-}
-
-// EqualLinkedResource returns true if the resource belongs to the same environment and application.
-func (b *BasicResourceProperties) EqualLinkedResource(prop *BasicResourceProperties) bool {
-	return strings.EqualFold(b.Application, prop.Application) && strings.EqualFold(b.Environment, prop.Environment)
-}
-
-type ResourceStatus struct {
-	OutputResources []outputresource.OutputResource `json:"outputResources,omitempty"`
-}
-
-func (in *ResourceStatus) DeepCopy(out *ResourceStatus) {
-	in.OutputResources = out.OutputResources
-}
-
-// OutputResource contains some internal fields like resources/dependencies that shouldn't be inlcuded in the user response
-func BuildExternalOutputResources(outputResources []outputresource.OutputResource) []map[string]interface{} {
-	var externalOutputResources []map[string]interface{}
-	for _, or := range outputResources {
-		externalOutput := map[string]interface{}{
-			"LocalID":  or.LocalID,
-			"Provider": or.ResourceType.Provider,
-			"Identity": or.Identity.Data,
-		}
-		externalOutputResources = append(externalOutputResources, externalOutput)
-	}
-
-	return externalOutputResources
 }
