+++ pkg/connectorrp/renderers/daprpubsubbrokers/renderer_test.go
@@ -17,6 +17,7 @@ import (
 	"github.com/project-radius/radius/pkg/connectorrp/renderers/dapr"
 	"github.com/project-radius/radius/pkg/kubernetes"
 	"github.com/project-radius/radius/pkg/resourcekinds"
+	"github.com/project-radius/radius/pkg/rp"
 	"github.com/project-radius/radius/pkg/rp/outputresource"
 	"github.com/stretchr/testify/require"
 	"gopkg.in/yaml.v2"
@@ -46,7 +47,7 @@ func Test_Render_Generic_Success(t *testing.T) {
 			Type: ResourceType,
 		},
 		Properties: datamodel.DaprPubSubBrokerProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: applicationID,
 				Environment: environmentID,
 			},
@@ -103,7 +104,7 @@ func Test_Render_Generic_MissingMetadata(t *testing.T) {
 			Type: ResourceType,
 		},
 		Properties: datamodel.DaprPubSubBrokerProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: applicationID,
 				Environment: environmentID,
 			},
@@ -130,7 +131,7 @@ func Test_Render_Generic_MissingType(t *testing.T) {
 			Type: ResourceType,
 		},
 		Properties: datamodel.DaprPubSubBrokerProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: applicationID,
 				Environment: environmentID,
 			},
@@ -159,7 +160,7 @@ func Test_Render_Generic_MissingVersion(t *testing.T) {
 			Type: ResourceType,
 		},
 		Properties: datamodel.DaprPubSubBrokerProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: applicationID,
 				Environment: environmentID,
 			},
@@ -231,7 +232,7 @@ func Test_Render_DaprPubSubAzureServiceBus_Success(t *testing.T) {
 			Type: ResourceType,
 		},
 		Properties: datamodel.DaprPubSubBrokerProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: applicationID,
 				Environment: environmentID,
 			},
@@ -276,7 +277,7 @@ func Test_Render_DaprPubSubMissingTopicName_Success(t *testing.T) {
 			Type: ResourceType,
 		},
 		Properties: datamodel.DaprPubSubBrokerProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: applicationID,
 				Environment: environmentID,
 			},
@@ -320,7 +321,7 @@ func Test_Render_DaprPubSubAzureServiceBus_InvalidResourceType(t *testing.T) {
 			Type: ResourceType,
 		},
 		Properties: datamodel.DaprPubSubBrokerProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: applicationID,
 				Environment: environmentID,
 			},
@@ -346,7 +347,7 @@ func Test_Render_InvalidApplicationID(t *testing.T) {
 			Type: ResourceType,
 		},
 		Properties: datamodel.DaprPubSubBrokerProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: "invalid-app-id",
 				Environment: environmentID,
 			},
@@ -376,7 +377,7 @@ func Test_Render_EmptyApplicationID(t *testing.T) {
 			Type: ResourceType,
 		},
 		Properties: datamodel.DaprPubSubBrokerProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Environment: environmentID,
 			},
 			Kind: resourcekinds.DaprGeneric,
