+++ pkg/connectorrp/renderers/daprstatestores/renderer_test.go
@@ -17,6 +17,7 @@ import (
 	"github.com/project-radius/radius/pkg/connectorrp/renderers"
 	"github.com/project-radius/radius/pkg/kubernetes"
 	"github.com/project-radius/radius/pkg/resourcekinds"
+	"github.com/project-radius/radius/pkg/rp"
 	"github.com/project-radius/radius/pkg/rp/outputresource"
 	"github.com/stretchr/testify/require"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -42,7 +43,7 @@ func Test_Render_Success(t *testing.T) {
 			Type: "Applications.Connector/daprStateStores",
 		},
 		Properties: datamodel.DaprStateStoreProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: applicationID,
 				Environment: environmentID,
 			},
@@ -85,7 +86,7 @@ func Test_Render_InvalidResourceType(t *testing.T) {
 			Type: "Applications.Connector/daprStateStores",
 		},
 		Properties: datamodel.DaprStateStoreProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: applicationID,
 				Environment: environmentID,
 			},
@@ -111,7 +112,7 @@ func Test_Render_SpecifiesUmanagedWithoutResource(t *testing.T) {
 			Type: "Applications.Connector/daprStateStores",
 		},
 		Properties: datamodel.DaprStateStoreProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: applicationID,
 				Environment: environmentID,
 			},
@@ -135,7 +136,7 @@ func Test_Render_UnsupportedKind(t *testing.T) {
 			Type: "Applications.Connector/daprStateStores",
 		},
 		Properties: datamodel.DaprStateStoreProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: applicationID,
 				Environment: environmentID,
 			},
@@ -161,7 +162,7 @@ func Test_Render_Generic_Success(t *testing.T) {
 			Type: "Applications.Connector/daprStateStores",
 		},
 		Properties: datamodel.DaprStateStoreProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: applicationID,
 				Environment: environmentID,
 			},
@@ -218,7 +219,7 @@ func Test_Render_Generic_MissingMetadata(t *testing.T) {
 			Type: "Applications.Connector/daprStateStores",
 		},
 		Properties: datamodel.DaprStateStoreProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: applicationID,
 				Environment: environmentID,
 			},
@@ -245,7 +246,7 @@ func Test_Render_Generic_MissingType(t *testing.T) {
 			Type: "Applications.Connector/daprStateStores",
 		},
 		Properties: datamodel.DaprStateStoreProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: applicationID,
 				Environment: environmentID,
 			},
@@ -274,7 +275,7 @@ func Test_Render_Generic_MissingVersion(t *testing.T) {
 			Type: "Applications.Connector/daprStateStores",
 		},
 		Properties: datamodel.DaprStateStoreProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: applicationID,
 				Environment: environmentID,
 			},
@@ -304,7 +305,7 @@ func Test_Render_InvalidApplicationID(t *testing.T) {
 			Type: "Applications.Connector/daprStateStores",
 		},
 		Properties: datamodel.DaprStateStoreProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: "invalid-app-id",
 				Environment: environmentID,
 			},
@@ -330,7 +331,7 @@ func Test_Render_EmptyApplicationID(t *testing.T) {
 			Type: "Applications.Connector/daprStateStores",
 		},
 		Properties: datamodel.DaprStateStoreProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Environment: environmentID,
 			},
 			Kind: datamodel.DaprStateStoreKindAzureTableStorage,
