+++ pkg/corerp/api/v20220315privatepreview/volume_conversion.go
@@ -9,6 +9,7 @@ import (
 	"github.com/project-radius/radius/pkg/armrpc/api/conv"
 	v1 "github.com/project-radius/radius/pkg/armrpc/api/v1"
 	"github.com/project-radius/radius/pkg/corerp/datamodel"
+	"github.com/project-radius/radius/pkg/rp"
 
 	azto "github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/go-autorest/autorest/to"
@@ -31,7 +32,7 @@ func (src *VolumeResource) ConvertTo() (conv.DataModelInterface, error) {
 			},
 		},
 		Properties: datamodel.VolumeResourceProperties{
-			BasicResourceProperties: v1.BasicResourceProperties{
+			BasicResourceProperties: rp.BasicResourceProperties{
 				Application: to.String(src.Properties.GetVolumeProperties().Application),
 			},
 			Kind: to.String(src.Properties.GetVolumeProperties().Kind),
@@ -86,7 +87,7 @@ func (dst *VolumeResource) ConvertFrom(src conv.DataModelInterface) error {
 		azProp := resource.Properties.AzureKeyVault
 		p := &AzureKeyVaultVolumeProperties{
 			Status: &ResourceStatus{
-				OutputResources: v1.BuildExternalOutputResources(resource.Properties.Status.OutputResources),
+				OutputResources: rp.BuildExternalOutputResources(resource.Properties.Status.OutputResources),
 			},
 			Kind:              azto.Ptr(resource.Properties.Kind),
 			Application:       azto.Ptr(resource.Properties.Application),
